+++ YUViewLib/src/parser/Mpeg2/AnnexBMpeg2.h
@@ -32,7 +32,7 @@
 
 #pragma once
 
-#include "AnnexB.h"
+#include "../AnnexB.h"
 
 namespace parser
 {
